webapi/webapi-vehicleinfo-ivi-tests/vehicleinfo/InteriorLightStatus_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset="utf-8">
<title>Vehicle Test: InteriorLightStatus - attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://rawgit.com/w3c/automotive-bg/master/data_spec.html">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var interiorLightStatus;

setup(function () {
  interiorLightStatus = vehicle.interiorLightStatus;
});

[
  ["boolean", "status", ""],
  ["object", "zone", "readonly"]
].forEach(function (attr) {
  var type = attr[0];
  var name = attr[1];
  var read = attr[2];

  test(function () {
    assert_true(name in interiorLightStatus, "the InteriorLightStatus." + name + " exists");
    assert_equals(typeof interiorLightStatus[name], type, "the type of InteriorLightStatus." + name);
    if ( read == "readonly" ) {
      assert_readonly(interiorLightStatus, name, "the InteriorLightStatus." + name + " is readonly");
    }
  }, "Check if " + read + " InteriorLightStatus." + name + " exists and type of " + type);
});

</script>
